requirements.txt
Babel==2.4.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
configparser==3.5.0
docutils==0.13.1
enum34==1.1.6
flake8==3.3.0
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
Pillow==4.2.1
py==1.4.34
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
Pygments==2.2.0
pytest==3.1.3
pytz==2017.2
requests==2.18.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
typing==3.6.1
urllib3==1.21.1
